ai_analyst/pyproject.toml
[project]
name = "ai-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.134.0",
    "google-genai>=1.65.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.41.0",
]

[project.scripts]
analyst = "main:run_a"
